% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructStudyToStudyAdjMat.R
\name{construct.study.to.study.adj.mat}
\alias{construct.study.to.study.adj.mat}
\title{Construct study-to-study adjacency matrix}
\usage{
construct.study.to.study.adj.mat(study.list, self.loop = FALSE)
}
\arguments{
\item{study.list}{a list of study accession phs ids}

\item{self.loop}{If FALSE, comparison of study to itself is set to 0}
}
\value{
matrix
}
\description{
Given a list of study accession phs ids of length n, return a n-by-n matrix
where the entry on i-th row j-th column is the number of times the i-th
study and the j-th study been requested by the same PIs
}
\examples{
construct.study.to.study.adj.mat(list("phs000501.v1.p1","phs000688.v1.p1","phs001516.v1.p1"))

}
