% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestDACTables.R
\name{dac.action.table.update}
\alias{dac.action.table.update}
\title{Updating the DAC action table}
\usage{
dac.action.table.update(
  update.to = format(Sys.Date(), "\%m/\%d/\%Y"),
  overwrite = TRUE,
  return.table = FALSE
)
}
\arguments{
\item{update.to}{String, date in"mm/dd/yyyy" format, by default it's the system date (today)}

\item{overwrite}{logical, whether the updated table should overwrite the previously saved table, TRUE by default}

\item{return.table}{logical, whether the function should return the updated table, FALSE by default}
}
\value{
(optionally) the updated table
}
\description{
Call this function to update the existing DAC action table data so it has the latest data
}
\examples{
\dontrun{
dac.action.table.update()
}

}
