% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateEveryTable.R
\name{dac.data.update.all}
\alias{dac.data.update.all}
\title{Update all DAC data}
\usage{
dac.data.update.all()
}
\description{
Updates every table that are locally stored in the package. Including
the nih_dac_actions_table and the all_nih_dac_studies_table
}
\examples{
\dontrun{
dac.data.update.all()
}

}
