% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARReviewTimelineSummary.R
\name{dar.review.timeline.summary}
\alias{dar.review.timeline.summary}
\title{DAR Review Timeline Summary}
\format{
A data frame with 6 variables:
\describe{
\item{DAC}{The name of the DAC (Data Access Committee)}
\item{AvgApprovalTime}{The average number of days for DARs from submitted by PI to be approved by the DAC}
\item{MedApprovalTime}{The median number of days for DARs from submitted by PI to be approved by the DAC}
\item{DARDailyAvg}{The average number of DAR submitted each day in the given timeframe}
\item{DARDailySD}{The standard deviation of DAR submitted each day in the given timeframe}
\item{DARTotal}{The total number of DAR submitted in the given timeframe}
}
}
\usage{
dar.review.timeline.summary(
  start.date,
  end.date,
  df = get.nih.dac.action.table()
)
}
\arguments{
\item{start.date}{String, date in "yyyy-mm-dd" format}

\item{end.date}{String, date in "yyyy-mm-dd" format}

\item{df}{optional, the dataframe to be analyzed, by default it is the nih_dac_action_table dataframe}
}
\value{
dataframe
}
\description{
Returns a summary dataframe of all DAC which had DAR submitted in the given timeframe
}
\examples{
\dontrun{
> dar.review.timeline.summary("2020-01-01","2020-10-15")
}

}
