% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-table-request-url.R
\name{formatTableRequestUrl}
\alias{formatTableRequestUrl}
\title{formatTableRequestUrl}
\usage{
formatTableRequestUrl(table.name, start.date, end.date)
}
\arguments{
\item{table.name}{String, name of table to query from}

\item{start.date}{String, date in mm/dd/yyyy format}

\item{end.date}{String, date in mm/dd/yyyy format}
}
\value{
the formated query string to be sent
}
\description{
Formats the given parameters as a query string that can be accepted by the dbGaP Data Access and Use Report
site (https://www.ncbi.nlm.nih.gov/projects/gap/cgi-bin/DataUseSummary.cgi)
table.name is one of table<1-9>, tablea1, tableb1, tablec1, tabled1
}
\examples{
formatTableRequestUrl('table2','01/01/2020','10/20/2020')
}
