% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_dac_action_table.R
\name{get.all.dac.action.table}
\alias{get.all.dac.action.table}
\title{Get all DAC action table}
\usage{
get.all.dac.action.table(start.date, end.date)
}
\arguments{
\item{start.date}{String, date in "mm/dd/yyyy" format}

\item{end.date}{String, date in "mm/dd/yyyy" format}
}
\value{
Dataframe containing all DAC actions performed within the specified time range
}
\description{
Get all DAC action table
}
\examples{
\dontrun{
get.all.dac.action.table('01/01/2015','12/31/2020')
}

}
