% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAllNIHDACStudiesTable.R
\name{get.all.nih.dac.studies.table}
\alias{get.all.nih.dac.studies.table}
\title{Returns The NIH DAC Studies Table}
\format{
A data frame with 5 variables:
\describe{
\item{DAC}{The DAC (Data Access Committee) in charge of the study}
\item{Study Accession}{The phs id of the study}
\item{Study Release Date}{The date in which the study was released}
\item{Embargo End Date}{The date in which the embargo of the study ends}
\item{Study Name}{The name of the study}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/projects/gap/cgi-bin/DataUseSummary.cgi?stDate=05\%2F13\%2F2020&endDate=11\%2F13\%2F2020&retTable=tablea1}
}
\usage{
get.all.nih.dac.studies.table()
}
\value{
Dataframe
}
\description{
Returns the locally stored dataframe containing all NIH DAC studies
}
\details{
This table contains all phs studies along with the DAC that they associated to and some of their
basic study information. This table is an exact copy of tablea1 on the Data Use Summary
Page.
}
