% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetApprovalTimeDiffTable.R
\name{get.approval.time.diff.table}
\alias{get.approval.time.diff.table}
\title{Get Approval time difference table}
\usage{
get.approval.time.diff.table(
  start.date,
  end.date,
  action.table.df = get.nih.dac.action.table()
)
}
\arguments{
\item{start.date}{String, date in 'yyyy-mm-dd' format}

\item{end.date}{String, date in 'yyyy-mm-dd' format}

\item{action.table.df}{Dataframe the time difference will be calculated on, by default it is the
locally stored nih dac action table}
}
\value{
dataframe
}
\description{
Returns a table with the approval time of all studies approved by dac in the given timeframe
}
\examples{
\dontrun{
table.with.new.columns <- get.approval.time.diff.table()
}
}
