% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGoogleScholarCitationTable.R
\name{get.google.scholar.citation.table}
\alias{get.google.scholar.citation.table}
\title{Get Google Scholar Citation Table}
\format{
A data frame with 3 variables:
\describe{
\item{phs_id}{phs id of the study}
\item{gs_citation_count}{number of search results returned when the phs_id is searched
without including version number (ex. phs000501.v1.p1 -> phs000501)}
\item{last_update}{The date in which the result was obtained on}
}
}
\usage{
get.google.scholar.citation.table()
}
\value{
dataframe
}
\description{
Retrieves the locally stored google scholar citation table.
}
\examples{
\dontrun{
gs.table <- get.google.scholar.citation.table()
>             phs_id gs_citation_count last_update
1    phs000501.v1.p1                 5  2020-12-07
2    phs000688.v1.p1                 9  2020-12-07
3    phs001516.v1.p1                 1  2020-12-07
4    phs001477.v1.p1                 0  2020-12-07
5    phs000710.v1.p1                 1  2020-12-07
6    phs000449.v2.p1                 2  2020-12-07
7    phs000678.v1.p1                 3  2020-12-07
8    phs000853.v1.p1                19  2020-12-07
}
}
