% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get.latest.approved.dar.date}
\alias{get.latest.approved.dar.date}
\title{Get Latest Approved Table}
\usage{
get.latest.approved.dar.date()
}
\value{
Date object representing the latest approved DAR date
}
\description{
Get the latest DAR approval date stored in nih_dac_action_table
}
\examples{
\dontrun{
get.latest.approved.dar.date()
> "2020-11-13"
}

}
