% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMonthlyStudyStatusTable.R
\name{get.monthly.study.status.table}
\alias{get.monthly.study.status.table}
\title{Get Monthly Study Status Table}
\format{
\describe{
\item{StudiesReleased}{The number of studies released that month in the given studies information dataframe}
\item{TotalRequests}{The number of requests made that made in the given DAC action information dataframe}
\item{StudiesReleasedCummulative}{The cummulative number of studies released up until that row}
\item{TotalRequestsCummulative}{The cummulative number of DAR made up until that row}
\item{StudiesReleasedGrowth}{The percentage growth of studies released compared to previous month}
\item{TotalRequestsGrowth}{The percentage growth of DAR compared to previous month}
}
}
\usage{
get.monthly.study.status.table(
  start.date,
  end.date,
  studies.df = all_nih_dac_studies_table,
  action.table.df = nih_dac_action_table
)
}
\arguments{
\item{start.date}{String, date in 'yyyy-mm-dd' format. Start date of the timeframe}

\item{end.date}{String, date in 'yyyy-mm-dd' format. End date of the timeframe}

\item{studies.df}{optional, dataframe. The dataframe containing studies information
to be used in the analysis, by default it is the all_nih_dac_studies_table dataframe}

\item{action.table.df}{optional, dataframe. The dataframe containing DAC action information
to be used in the analysis, by default it is the nih_dac_action_table dataframe}
}
\value{
a dataframe
}
\description{
Returns a table comparing study released and requests made each month
}
\examples{
\dontrun{
> get.monthly.study.status('2020-01-01','2020-12-31')
        Month StudiesReleased TotalRequests StudiesReleasedCummulative TotalRequestsCummulative ...
1  2020-01-01              17          1514                         17                     1514 ...
2  2020-02-01              29          2837                         46                     4351 ...
3  2020-03-01              26          2696                         72                     7047 ...
4  2020-04-01              30          3261                        102                    10308 ...
5  2020-05-01              21          2922                        123                    13230 ...
6  2020-06-01              11          3716                        134                    16946 ...
7  2020-07-01              16          3144                        150                    20090 ...
8  2020-08-01               8          3131                        158                    23221 ...
}


}
