% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPhsStudiesTable.R
\name{get.phs.studies.table}
\alias{get.phs.studies.table}
\title{Get phs Studies Table}
\format{
A data frame with 7 variables:
\describe{
\item{Consent group}{Ways in which the data can be used for}
\item{Is IRBrequired?}{Whether IRB permission is needed to use the data}
\item{Data Access Committee}{The DAC to contact for access to the study}
\item{phs_id}{The phs_id associated to the study}
\item{study_key}{The internal key of the study, currently used only for web scraping purposes}
\item{PI}{The principle investigators responsible for the study. Separated by semicolon}
\item{Number ofparticipants}{The number of participants included in the study}
}
}
\usage{
get.phs.studies.table()
}
\value{
dataframe
}
\description{
This table contains some of the meta information about each studies that are listed
under the all_nih_dac_actions table. The data is obtained through scraping the content
on dbGAP
}
\examples{
\dontrun{
table <- get.phs.studies.table()
}

}
