% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPhsStudyTermFrequencyTable.R
\name{get.phs.study.term.frequency.table}
\alias{get.phs.study.term.frequency.table}
\title{Get Phs study term frequency table}
\usage{
get.phs.study.term.frequency.table(phs.id)
}
\arguments{
\item{phs.id}{the phs id of the study}
}
\value{
dataframe
}
\description{
Look up all research statements associated to the given phs id and return a
table with words appeared in the research statements along with their frequency
}
\details{
Reference: http://www.sthda.com/english/wiki/text-mining-and-word-cloud-fundamentals-in-r-5-simple-steps-you-should-know
}
\examples{
\dontrun{
term.frequency.table <- get.phs.study.term.frequency.table('phs000688.v1.p1')
}

}
