% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPITable.R
\name{get.pi.table}
\alias{get.pi.table}
\title{Get PI table}
\format{
A dataframe with 7 variables
\describe{
\item{PI}{The ID of the PI}
\item{DAR}{The number of DAR submitted by the PI within the timeframe}
\item{DAC}{The number of DACs the PI submitted DAR to within the timeframe}
\item{Project}{The number of projects the PI submitted DAR to within the timeframe}
\item{Approved by DAC}{The number of DAR submitted by the PI that are approved wiuthin the timeframe}
\item{Rejected by DAC}{The number of DAR submitted by the PI that are rejected within the timeframe}
\item{Data downloaded}{The number of datasets that the PI downloaded within the timeframe}
}
}
\usage{
get.pi.table(start.date, end.date, action.table.df = nih_dac_action_table)
}
\arguments{
\item{start.date}{String, date in 'yyyy-mm-dd' format. The start of the timeframe}

\item{end.date}{String, date in 'yyyy-mm-dd' format. The end of the timeframe}

\item{action.table.df}{Optional, Dataframe. The table to perform analysis on, defaults to be the nih_dac_action_table.}
}
\value{
dataframe
}
\description{
Returns a table about PI (Principal Investigator) and the status of their DARs
}
\examples{
\dontrun{
> get.pi.table("2020-01-01","2020-12-31")
    PI DAR DAC Project Approved by DAC Rejected by DAC Data downloaded
1   10008   1   1       1               1               0               0
2   10024   1   1       1               1               0               1
3   10035   1   1       1               1               0               0
4   10040   8   3       2               8               0               8
5   10046   5   2       3               4               1               1
6   10047   1   1       1               1               0               0
7    1006   5   1       1               5               5               3
8   10067   2   2       1               2               0               0
9   10074   4   1       4               4               2               0
10  10078   4   2       1               4               0               3
}



}
