% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMonthlyStudyGrowth.R
\name{get.study.momthly.growth}
\alias{get.study.momthly.growth}
\title{Get Study Monthly Growth}
\usage{
get.study.momthly.growth(phs.id)
}
\arguments{
\item{phs.id}{String, the phs id of the study}
}
\value{
dataframe
}
\description{
Returns a data frame with month and the number of times the study is requested per Month
}
\examples{
\dontrun{
get.study.momthly.growth('phs001542.v1.p1')
>       Month DAR
1  2019-08-01   1
2  2019-09-01   1
3  2019-10-01   2
4  2019-11-01   1
5  2019-12-01   2
6  2020-02-01   2
7  2020-03-01   1
8  2020-05-01   2
9  2020-06-01   4
10 2020-08-01   5
11 2020-09-01   6
12 2020-10-01   4
13 2020-11-01   1

}
}
