% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStudySummaryTable.R
\name{get.study.summary.table}
\alias{get.study.summary.table}
\title{Get Study Summary Table}
\format{
A data frame with 12 variables:
\describe{
\item{StudyAccession}{The phs number of the study}
\item{TotalRequest}{Total number of DAR submitted by PI for the study in the given timeframe}
\item{TotalApproval}{Total number of DAR approved by DAC for the study in the given timeframe}
\item{TotalRejected}{Total number of DAR rejected by DAC for the study in the given timeframe}
\item{TotalDownloaded}{Total number of times the study is downloaded}
\item{AvgApprovalTime}{The average number of days it takes for a DAR for the study to be approved}
\item{ApprovalRate}{Proportion of DAR that are approved. Calcualted by TotalApproved / (TotalApproved+TotalRejected)}
\item{DownloadRate}{Proportion of DAR that are downloaded. Calculated by TotalDownload / TotalRequest}
\item{DAC}{Name of the DAC that submitted the DAR for the study}
\item{Study Release Date}{Date in which the study is released}
\item{Embargo End Date}{Date in which the embargo of the study is lifted}
\item{Study Name}{Name of the study}
}
}
\usage{
get.study.summary.table(
  start.date = "2000-01-01",
  end.date = format(Sys.Date()),
  df = nih_dac_action_table
)
}
\arguments{
\item{start.date}{String, date in "yyyy-mm-dd" format}

\item{end.date}{String, date in "yyyy-mm-dd" format}

\item{df}{optional, the dataframe to be analyzed, by default it is the nih_dac_action_table dataframe}
}
\value{
dataframe
}
\description{
Returns a summary dataframe of all studies made in the given timeframe. Note that some studies
can be both approved and rejected (ex. approved but later rejected ), or neither (still in process of approval)
}
\examples{
\dontrun{
> get.study.summary.table("2020-01-01","2020-10-15")
       StudyAccesion TotalRequest TotalApproved TotalRejected TotalDownload AvgApprovalTime ...
1    phs000001.v3.p1          106            94             5            59       31.205245 ...
2  phs000007.v31.p12          440           312           133           228       17.452357 ...
3    phs000016.v2.p2           27            24             0             9       21.924392 ...
4    phs000017.v3.p1          188           171             0           108       22.340318 ...
5    phs000018.v2.p1           16            12             0             7       57.746181 ...
6    phs000019.v1.p1           45            36             8            16       11.507002 ...
7    phs000020.v2.p1           51            49             0            22       21.797038 ...
8    phs000021.v3.p2          196           184             0            99       21.470720 ...
9    phs000048.v1.p1           34            30             5            15        7.769923 ...
10   phs000086.v3.p1           37            31             0            20       49.872446 ...
......
}

}
