% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructStudyToStudyAdjMat.R
\name{get.study.to.study.adj.mat}
\alias{get.study.to.study.adj.mat}
\title{Get Study to Study Adjacency Matrix}
\usage{
get.study.to.study.adj.mat()
}
\value{
matrix
}
\description{
Returns the locally stored study-to-study adjacency matrix. See documentation
on construct.study.to.study.adj.mat for detail.
}
\examples{
\dontrun{
study.to.study.adj.df <- get.study.to.study.adj.mat()
}
}
