% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIHDACActionTable.R
\docType{data}
\name{nih_dac_action_table}
\alias{nih_dac_action_table}
\title{The NIH DAC Action Table}
\format{
A data frame with 11 variables:
\describe{
\item{DAC}{The DAC (Data Access Committee) in which the DAR was from}
\item{PI}{ID of the PI who made the request}
\item{Project}{ID of the project in which the request was made}
\item{DAR}{The ID of the request}
\item{Study accesion}{The phs number of the study in which the DAR was made for}
\item{Submitted by PI}{The time when the DAR was made}
\item{Approved by SO}{The time when the DAR is approved by SO}
\item{Approved by DAC}{The time when the DAR is approved by DAC}
\item{Rejected by DAC}{The time when the DAR is rejected by DAC}
\item{Revision requested by DAC}{The time when the DAR is requested a revision by DAC}
\item{Data downloaded}{Whether the data requested is downloaded}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/projects/gap/cgi-bin/DataUseSummary.cgi?DAC=all&actType=all&stDate=10/13/2020&endDate=10/23/2020}
}
\usage{
nih_dac_action_table
}
\description{
This table contains all DAR (Data Access Requests) that are approved within a period of time
}
\keyword{datasets}
