% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructStudyToStudyAdjMat.R
\name{pca.plot.studies}
\alias{pca.plot.studies}
\title{PCA Plot of network adjacency matrix}
\usage{
pca.plot.studies(adj.mat = get.study.to.study.adj.mat())
}
\arguments{
\item{adj.mat}{The adjacency matrix to be used to plot, By default the study to study adjacency matrix}
}
\value{
ggplot figure
}
\description{
This function is currently EXPERIMENTAL and the method is not vigorously verified.
Any conclusion drawn from the plotted figure should be treated with caution.
}
\details{
Given the adjacency matrix of study to study matrix, where edge weight indicates
the number of PIs who requested the two studies together, plot a PCA-reduced version
of the adjancency matrix in 2 dimension, colored by DAC.

The following operations are preformed to produce the plot: The adjacency matrix
is first min-max normalized column-wise, NAs are filled by 0, then the adjancency
matrix is reduced to 2 dimension.
}
\examples{
\dontrun{
pca.plot.studies()
}
}
