% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestDACTables.R
\name{request.all.dac.action.table}
\alias{request.all.dac.action.table}
\title{Get all DAC action table}
\usage{
request.all.dac.action.table(start.date, end.date)
}
\arguments{
\item{start.date}{String, date in "mm/dd/yyyy" format}

\item{end.date}{String, date in "mm/dd/yyyy" format}
}
\value{
Dataframe containing all DAC actions performed within the specified time range
}
\description{
Get all DAC action table
}
\examples{
\dontrun{
> request.all.dac.action.table('09/01/2020','10/01/2020')
     DAC    PI Project       DAR  Study accesion  Submitted by PI   ... Data downloaded
2    CDAC   563     501  65316.v4 phs000688.v1.p1 09/07/2020 07:22  ... no
3    CDAC   563   11888  74027.v3 phs000688.v1.p1 09/09/2020 09:31  ... no
4    CDAC  1280     849  73340.v3 phs000688.v1.p1 08/28/2020 11:46  ... yes in previous version
5    CDAC  2418    1395  85271.v3 phs000688.v1.p1 09/02/2020 14:55  ... no
6    CDAC  3464    1955  48719.v9 phs000688.v1.p1 09/21/2020 15:05  ... no
......
}

}
