% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestDACTables.R
\name{request.all.nih.dac.studies.table}
\alias{request.all.nih.dac.studies.table}
\title{Request all nih dac studies table (tablea1)}
\usage{
request.all.nih.dac.studies.table(
  start.date = "01/01/2000",
  end.date = format(Sys.Date(), "\%m/\%d/\%Y")
)
}
\arguments{
\item{start.date}{String, date in 'mm/dd/yyyy' format}

\item{end.date}{String, date in 'mm/dd/yyyy' format}
}
\value{
dataframe
}
\description{
Scrapes tablea1, see get.all.nih.dac.studies.table for detail
}
\examples{
\dontrun{
tablea1 <- request.all.nih.dac.studies.table()
}
}
