% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestGoogleScholarCitationTable.R
\name{request.google.scholar.citation.table}
\alias{request.google.scholar.citation.table}
\title{Request Google Scholar Citation Table}
\usage{
request.google.scholar.citation.table(phs.id.list, wait.for = 2)
}
\arguments{
\item{phs.id.list}{list of phs.id}

\item{wait.for}{integer, seconds to wait for before sending another request}
}
\value{
dataframe
}
\description{
Given list of phs id, create a dataframe with their corresponding number
of citations (number of results returned from searching) on google scholar
Empirically, setting the wait time >= 2 prevents captcha from showing up
For table column see documentation for get.google.scholar.citation.table()
}
\examples{
\dontrun{
df <- get.nih.dac.action.table()
unique.study.ids <- unique(df['Study accesion'])$`Study accesion`
gs.df <- request.google.scholar.citation.table(unique.study.ids,2)
}
}
