% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestGoogleScholarCitationTable.R
\name{request.google.scholar.result.count}
\alias{request.google.scholar.result.count}
\title{Request Google Scholar Result Count}
\usage{
request.google.scholar.result.count(term, start.year = "", end.year = "")
}
\arguments{
\item{term}{String, the search term for google scholar}

\item{start.year}{String, year in 'yyyy' format. The earliest year in which
the result should contain}

\item{end.year}{String, year in 'yyyy' format. The latest year in which the
result should contain}
}
\value{
integer
}
\description{
Request Google Scholar Result Count
}
\examples{
\dontrun{
# The number of citations for phs000449
study.gs.citation.count <- request.google.scholar.result.count('phs000449')
}
}
