% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructStudyToStudyAdjMat.R
\name{save.study.to.study.adj.mat}
\alias{save.study.to.study.adj.mat}
\title{Save Study to Study Adjacency Matrix}
\usage{
save.study.to.study.adj.mat()
}
\description{
Saves the study-to-study adjacency matrix. See documentation
on construct.study.to.study.adj.mat for detail.
}
\examples{
\dontrun{
save.study.to.study.adj.mat()
}
}
