% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateEveryTable.R
\name{update.every.table}
\alias{update.every.table}
\title{Update Every Table}
\usage{
\method{update}{every.table}()
}
\description{
Updates every table that are locally stored in the package. Including
the nih_dac_actions_table and the all_nih_dac_studies_table
}
\examples{
\dontrun{
update.every.table
}

}
