% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.dac.report.R
\name{compile.dac.report}
\alias{compile.dac.report}
\title{Compile DAC Report}
\usage{
compile.dac.report(dac, author, start.date, end.date, ...)
}
\arguments{
\item{dac}{String, the name of the DAC from which the report will be generated for, one of
"CDAC", "ES DAC", "JAAMH", "Kids First DAC", "NCATS", "NCI DAC", "NEI", "NHGRI",
"NHLBI", "NIAID", "NIAMS", "NICHD", "NIDCD DAC", "NIDCR", "NIDDK", "NIGMS", "NINDS"
"NINR DAC". For a list of all currently supported DAC names call "CDAC", "ES DAC", "JAAMH", "Kids First DAC", "NCATS", "NCI DAC", "NEI", "NHGRI", "NHLBI", "NIAID", "NIAMS", "NICHD", "NIDCD DAC", "NIDCR", "NIDDK", "NIGMS", "NINDS", "NINR DAC"}

\item{author}{String, The name of the author who created this report}

\item{start.date}{String, date in "yyyy-mm-dd" format, the start date of the timeframe
in which the analysis will be based on}

\item{end.date}{String, date in "yyyy-mm-dd" format, the end date of the timeframe in
which the analysis will be based on}

\item{...}{extra arguments to be provided to rmarkdown::render}
}
\description{
Creates a report on the DAR (Data Access Request) status of the given DAC
(Data Access Commitee) and compare it to the rest of the NIH institutes. The
generated report will be a word document. "CDAC", "ES DAC", "JAAMH", "Kids First DAC", "NCATS", "NCI DAC", "NEI", "NHGRI", "NHLBI", "NIAID", "NIAMS", "NICHD", "NIDCD DAC", "NIDCR", "NIDDK", "NIGMS", "NINDS", "NINR DAC"
}
\examples{
\dontrun{
compile.dac.report('NIAID','Hoyin Chu, Christopher Marcum', '2015-01-01', '2019-12-31')
}

}
