% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DACDataUpdateAll.R
\name{dac.data.update.all}
\alias{dac.data.update.all}
\title{Update all DAC data}
\usage{
dac.data.update.all(
  update.to = format(Sys.Date(), "\%m/\%d/\%Y"),
  wait.for = 2
)
}
\description{
Updates every table that are locally stored in the package. Including
the nih_dac_actions_table and the all_nih_dac_studies_table
}
\examples{
\dontrun{
dac.data.update.all()
}

}
