% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefreshLocalData.R
\name{refresh.local.data}
\alias{refresh.local.data}
\title{Refresh Local Data}
\usage{
refresh.local.data(start.date = as.Date("2000-01-01"), end.date = Sys.Date())
}
\arguments{
\item{start.date}{the earliest date to get data from. Defaults to be 2000-01-01}

\item{end.date}{the latest date to get data from. Default to be the system date}
}
\description{
Request all data and overwrite the existing locally
stored data. May take 1-2 hours
}
\examples{
\dontrun{
refresh.local.data()
}

}
